fix(state): isolate registry seam ownership
(cherry picked from commit b57710cb70e6a80b60b6b3ebae44ffe8289850a1)
