+++ src/lib/state/registry.ts
@@ -7,24 +7,22 @@ import {
   inferenceSelectionRegistryFields,
   normalizeInferenceSelection,
 } from "../inference/selection";
-import { normalizeToolDisclosure, type ToolDisclosure } from "../tool-disclosure";
+import { normalizeToolDisclosure } from "../tool-disclosure";
 import {
   applyAddExtraProvider,
   applyRemoveExtraProvider,
   isValidExtraProviderName,
   readExtraProviders,
 } from "./extra-providers";
-import type { OpenClawImagePluginInstall } from "./openclaw-plugin-restore";
-import { normalizeSandboxMcpState, type SandboxMcpState } from "./registry-mcp";
-import type { SandboxMessagingState } from "./registry-messaging";
+import { withLock } from "./registry/lock";
+import { load, save } from "./registry/persistence";
+import { normalizeSandboxMcpState } from "./registry-mcp";
 import {
   normalizeBaselineExclusions,
   normalizeBaselineExclusionTransition,
   retainedDefaultSandbox,
 } from "./registry-normalization";
 import * as reversibleRemoval from "./registry-reversible-removal";
-import { withLock } from "./registry/lock";
-import { load, save } from "./registry/persistence";
 
 export {
   getSandboxEntryDisplayInference,
@@ -33,11 +31,13 @@ export {
   type SandboxEntryInference,
 } from "./registry-entry-view";
 
-import type { WebSearchProvider } from "../inference/web-search";
-import {
-  type DcodeAutoApprovalMode,
-  isDcodeAutoApprovalMode,
-} from "../onboard/dcode-auto-approval";
+import { isDcodeAutoApprovalMode } from "../onboard/dcode-auto-approval";
+import type {
+  BaselineExclusionEntry,
+  BaselineExclusionTransition,
+  CustomPolicyEntry,
+  SandboxEntry,
+} from "./registry/types";
 import {
   cloneSandboxMessagingState,
   getConfiguredMessagingChannels as getRegistryConfiguredMessagingChannels,
@@ -45,8 +45,9 @@ import {
   setChannelDisabled as setRegistryChannelDisabled,
 } from "./registry-messaging";
 
-export type { McpBridgeEntry, SandboxMcpState } from "./registry-mcp";
-
+// Compatibility exports for #7694. The registry/types, registry/lock, and
+// registry/persistence modules are authoritative. New code must import those
+// modules directly. Remove these exports after facade callers migrate.
 export {
   acquireLock,
   classifyExistingLock,
@@ -55,12 +56,22 @@ export {
   LOCK_OWNER,
   LOCK_RETRY_MS,
   LOCK_STALE_MS,
-  releaseLock,
   type RegistryLockDecision,
+  releaseLock,
   withLock,
 } from "./registry/lock";
-
 export { load, REGISTRY_FILE, save } from "./registry/persistence";
+export type {
+  BaselineExclusionEntry,
+  BaselineExclusionTransition,
+  BaselineExclusionTransitionOperation,
+  CustomPolicyEntry,
+  SandboxEntry,
+  SandboxGpuProofResult,
+  SandboxGpuProofStatus,
+  SandboxRegistry,
+} from "./registry/types";
+export type { McpBridgeEntry, SandboxMcpState } from "./registry-mcp";
 
 export {
   getConfiguredMessagingChannelsFromEntry,
@@ -69,145 +80,6 @@ export {
   getMessagingPlanFromEntry,
   type SandboxMessagingState,
 } from "./registry-messaging";
-
-export interface CustomPolicyEntry {
-  name: string;
-  content: string;
-  /** Desired content reserved before a crash-safe generated-policy transition. */
-  pendingContent?: string;
-  sourcePath?: string;
-  appliedAt?: string;
-}
-
-export interface BaselineExclusionEntry {
-  /** Persistence schema version for this reviewed exclusion intent. */
-  version: 1;
-  /** Agent baseline that supplied the reviewed entry. */
-  agent: string;
-  /** Exact baseline network policy key excluded, e.g. "nous_research". */
-  key: string;
-  /** Digest of the reviewed baseline entry content the approval was bound to. */
-  digest: string;
-  /** When the exclusion was acknowledged. */
-  acknowledgedAt?: string;
-  /** Agent build/version recorded when the exclusion was last applied. */
-  appliedAgentVersion?: string | null;
-}
-
-export type BaselineExclusionTransitionOperation = "exclude" | "restore";
-
-/**
- * Durable journal for the one cross-system baseline mutation that is in flight.
- * `baselineExclusions` remains the last committed operator intent until this
- * transaction is published after the live OpenShell mutation succeeds.
- */
-export interface BaselineExclusionTransition {
-  id: string;
-  operation: BaselineExclusionTransitionOperation;
-  exclusion: BaselineExclusionEntry;
-  /** Exact live-entry digest that completes the transition; null means absent. */
-  targetLiveDigest: string | null;
-  startedAt: string;
-}
-
-// Outcome of the last live sandbox GPU proof run during onboarding/recovery.
-// `status` separates a configured-but-unverified GPU from one whose CUDA
-// usability was actually proven (`verified`) or actively failed a live proof
-// (`failed`, e.g. Jetson `/dev/nvmap` permission errors). Persisted so
-// `nemoclaw <sandbox> status` can report proof state instead of treating any
-// configured GPU as healthy (#4231).
-export type SandboxGpuProofStatus = "verified" | "unverified" | "failed";
-
-export interface SandboxGpuProofResult {
-  status: SandboxGpuProofStatus;
-  // True only when a CUDA-usability proof (cuInit via libcuda) actually passed.
-  cudaVerified: boolean;
-  // Label of the last proof that determined `status`.
-  label?: string | null;
-  // Redacted, truncated diagnostic captured when the proof failed.
-  detail?: string | null;
-  at: string;
-}
-
-export interface SandboxEntry extends Partial<InferenceSelection> {
-  name: string;
-  /** Route-only placeholder created before sandbox creation; never eligible as the default. */
-  pendingRouteReservation?: true;
-  /** Onboard session that owns a pending reservation, so resume preserves its own row while abandoned reservations stay reconcilable. */
-  reservationSessionId?: string;
-  createdAt?: string;
-  gpuEnabled?: boolean;
-  hostGpuDetected?: boolean;
-  sandboxGpuEnabled?: boolean;
-  sandboxGpuMode?: "auto" | "1" | "0" | string | null;
-  sandboxGpuDevice?: string | null;
-  sandboxGpuProof?: SandboxGpuProofResult | null;
-  openshellDriver?: string | null;
-  openshellVersion?: string | null;
-  policies?: string[];
-  customPolicies?: CustomPolicyEntry[];
-  /** Operator exclusions from the agent baseline policy, replayed on rebuild. */
-  baselineExclusions?: BaselineExclusionEntry[];
-  /** Crash-recoverable journal for an exclusion/restore live-policy mutation. */
-  baselineExclusionTransition?: BaselineExclusionTransition;
-  policyTier?: string | null;
-  // True once the onboard policy step has fully completed and reconciled the
-  // effective preset selection (set by the post-policy registry write). Absent
-  // on a sandbox whose registration recorded only boot-time presets but whose
-  // policy step never finished — so re-onboard knows whether `policies`
-  // represents a final selection it can carry forward. See #4621.
-  policyPresetsFinalized?: boolean;
-  webSearchEnabled?: boolean;
-  /** Selected disclosure preference; model compatibility safeguards may downgrade runtime behavior. */
-  toolDisclosure?: ToolDisclosure;
-  /** Enables backend-neutral trace export to the fixed local OTLP collector boundary. */
-  observabilityEnabled?: boolean;
-  /** Image-baked permission to expose DCode's per-thread auto-approval opt-in. */
-  dcodeAutoApprovalMode?: DcodeAutoApprovalMode;
-  /** Durable provider identity for enabled managed web search. */
-  webSearchProvider?: WebSearchProvider | null;
-  agent?: string | null;
-  agentVersion?: string | null;
-  /** Plugin install baseline captured before state is restored into a fresh OpenClaw image. */
-  openclawImagePluginInstalls?: OpenClawImagePluginInstall[];
-  // NemoClaw build fingerprint (the NemoClaw CLI/build version) stamped only on
-  // NemoClaw-managed images at create/rebuild time. `upgrade-sandboxes` compares
-  // it against the running NemoClaw build so an image/build change with an
-  // unchanged agent version is still detected as needing a rebuild. Custom-image
-  // (`--from`) sandboxes are intentionally left without a fingerprint so they
-  // are never auto-rebuilt onto the default image (#5026).
-  nemoclawVersion?: string | null;
-  fromDockerfile?: string | null;
-  hermesAuthMethod?: "oauth" | "api_key" | null;
-  imageTag?: string | null;
-  messaging?: SandboxMessagingState;
-  mcp?: SandboxMcpState;
-  hermesToolGateways?: string[];
-  hermesDashboardEnabled?: boolean;
-  hermesDashboardPort?: number | null;
-  hermesDashboardInternalPort?: number | null;
-  hermesDashboardTui?: boolean;
-  dashboardPort?: number | null;
-  /** Remote dashboard exposure was included in the sandbox's generated config. */
-  dashboardRemoteBindPrepared?: boolean;
-  /** Generation proving which durable same-name recreate registered this row. */
-  lifecycleGeneration?: string;
-  /** Hashed OpenShell identity paired with lifecycleGeneration for exact recovery. */
-  lifecycleLiveIdentityFingerprint?: string;
-  // OpenShell gateway registration name and host port bound to this sandbox.
-  // Persisted so later lifecycle commands operate on the sandbox's own gateway
-  // instead of the process-global `nemoclaw` singleton — a second sandbox on a
-  // different NEMOCLAW_GATEWAY_PORT no longer recreates/kills the first (#4422).
-  gatewayName?: string | null;
-  gatewayPort?: number | null;
-}
-
-export interface SandboxRegistry {
-  sandboxes: Record<string, SandboxEntry>;
-  defaultSandbox: string | null;
-  defaultSelectionRevision?: number;
-  extraProviders?: string[];
-}
 
 export type SandboxRemovalReceipt = reversibleRemoval.RegistryRemovalReceipt<SandboxEntry>;
